Add files via upload
Day 5 Assignment contains below questions:
1. Do the below programs in anonymous function & IIFE 
a. Print odd numbers in an array 
b. Convert all the strings to title caps in a string array 
c. Sum of all numbers in an array d. Return all the prime numbers in an array 
e. Return all the palindromes in an array 
f. Return median of two sorted arrays of the same size 
g. Remove duplicates from an array h. Rotate an array by k times

3. Do the below programs in arrow functions 
a. Print odd numbers in an array 
b. Convert all the strings to title caps in a string array 
c. Sum of all numbers in an array 
d. Return all the prime numbers in an array 
e. Return all the palindromes in an array